lambda_functions/authorizer/requirements.txt
python-jose[cryptography]==3.3.0
aws-xray-sdk==2.12.1
